Player saving is now online.  Lots more to do for a complete database system, even just for the player database, but progress is being made.
Note:  The database schema is fully managed in code.  No changes should be manually
made to the database schema.  Add a block to the function in database.cpp.

Note: Most database schema changes are not safe to be made while the game server is
running.  Reboot the game server and allow it to update the schema itself.

BKrev: 3d6e9ad5Or6TYXje4oRblwwYA7L8mA