Lots of stuff done with this batch. As of now, all of the origional Network and character code has been ripped out and replaced with a much cleaner interface.  The new stuff will make it pretty easy to build in the editor support and online creation system as well as everything else that goes into this platform.
Main command tree searches now happen in the main thread instead of
in the command executor thread pool, but this should not be a problem.

Commands no longer receive the string that caused it to be executed as a
parameter.  It would be silly to pollute the nice structure that we
have built with command classes that change their operation based on
what string was used to call them.

BKrev: 3dbf8014TjiHpA1xa90gemKGs77_Tw